Stop reversing at low throttle + turn
When driving with just a tiny bit of throttle (barely any at all), and turning, it would actually allow the robot to drive backwards due to the slow side being simply subtracted off.  This edit forces the fast side to be the same magnitude as the slow side if it woudl otherwise be less.  This will mean more stick combinations will spin, but might feel more predictable to drivers.  Will test more to see.